requirements.txt
discord.py
discord.py[voice]
pytube
python-dotenv
youtube-search
validators
